locale.c: Quit no if strftime() fails with EINVAL
It turns out that some platforms set errno to EINVAL when the libc
function strftime() finds the input format to be in error, though this
is not specified in the Standard (a defect there in my opinion).
Windows in particular does set errno.  Quit when an EINVAL errno is
detected.  On platforms that don't set EINVAL, the current behavior is
retained where the loop is executed again and again to rule out that the
error isn't a buffer size problem.