locale.c: Avoid unnecessary locale changes
In our emulation of nl_langinfo(), prior to this commit, we always
toggled the locales of two categories to the necessary ones.  Except
that there are a number of paths through the code that don't require
this toggling.  This commit narrows the scope of the toggling to just
those cases that need it.