Return C values if localeconv() not available
localeconv() is a C89 function.  But the existing code handled the case
if it isn't available by returning an empty hash if POSIX::localeconv()
is called.  Prior commits have made it actually easier to instead return
the hash populated with the values that the C locale has, which is more
desirable than the current state.  And that is what this commit does.

There are reasons that localeconv() might not be available.  It still is
broken on MingW, and until recently on older Windows releases that we
now no longer support.  Workarounds were developed for those cases, but
there may be other platforms where it ends up broken without a
workaround, and this code handles them better and more simply than
before this commit.